circle dep fix in select
