+++ packages/ui/components/form/select/Select.tsx
@@ -1,6 +1,6 @@
 import { useId } from "@radix-ui/react-id";
 import * as React from "react";
-import type { GroupBase, Props, SingleValue, MultiValue } from "react-select";
+import type { GroupBase, SingleValue, MultiValue } from "react-select";
 import ReactSelect from "react-select";
 
 import { useLocale } from "@calcom/lib/hooks/useLocale";
@@ -9,17 +9,7 @@ import cx from "@calcom/ui/classNames";
 import { Label } from "../inputs/Label";
 import { inputStyles } from "../inputs/TextField";
 import { getReactSelectProps } from "./selectTheme";
-
-export type SelectProps<
-  Option,
-  IsMulti extends boolean = false,
-  Group extends GroupBase<Option> = GroupBase<Option>
-> = Props<Option, IsMulti, Group> & {
-  variant?: "default" | "checkbox";
-  "data-testid"?: string;
-  size?: "sm" | "md";
-  grow?: boolean;
-};
+import type { SelectProps } from "./types";
 
 export const Select = <
   Option,
